requirements.txt
python-telegram-bot==13.7
yt-dlp
python-dotenv
requests
aiogram
